refactor: DTO들이 모두 record로 구현되도록 수정
