+++ backend/src/main/java/com/now/naaga/game/application/GameService.java
@@ -1,6 +1,7 @@
 package com.now.naaga.game.application;
 
 import com.now.naaga.game.application.dto.CreateGameCommand;
+import com.now.naaga.game.application.dto.FindGameByIdCommand;
 import com.now.naaga.game.application.dto.FindGameByStatusCommand;
 import com.now.naaga.game.application.dto.FinishGameCommand;
 import com.now.naaga.game.domain.Game;
