+++ backend/src/main/java/com/now/naaga/member/application/dto/MemberCommand.java
@@ -1,25 +1,5 @@
 package com.now.naaga.member.application.dto;
 
-public class MemberCommand {
-
-    private String email;
-
-    private String password;
-
-    private MemberCommand() {
-    }
-
-    public MemberCommand(final String email,
-                         final String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
+public record MemberCommand(String email,
+                            String password) {
 }
